EER Diagram Added
The most notable change made to the EER diagram was the removal of the
entity DonationType. I figured we would just have an attribute that
would be a dropdown menu in donations.

I did not add the entity that Prof. Marta was discussing in class about
a campaign having multiple parts. Instead of making the assumption on
whether a campaign does have multiple parts and whether that needs to
be recorded, we should instead verify any of those assumptions.