Worked on Updates
I was able to pull data from the database onto the text fields and combo
boxes, unfortunately when you go to select donor you can only see item
1-4 but when you click on it it displays the information but the info
seems a little off. I wasn't able to get to the update aspect of my part
as I was learning as I was going through the code.